requirements.txt
nltk
newspaper3k
news-please
Unidecode
pymongo
# colorama